+++ src/main.ts
@@ -5,6 +5,7 @@ import router from "@/router/router";
 import App from "@/App.vue";
 import i18n from "@/i18n/i18n";
 import { formkit, formkitConfig } from "@/formkit";
+import highlight from "@/highlight";
 import { FontAwesomeIcon } from "@/fontawesome";
 import "@/index.css";
 
@@ -15,6 +16,7 @@ const app = createApp(App) //
   .use(pinia)
   .use(i18n)
   .use(formkit, formkitConfig)
+  .use(highlight)
   .component("icon", FontAwesomeIcon);
 
 // Use the Matomo plugin only if configured in env.