+++ src/metadata/CreateMetadata.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts">
 import { useRouter } from "vue-router";
+import { FormKit } from "@formkit/vue";
 import { useAuth } from "@/auth/auth.composable";
 import useMinkBackend from "@/api/backend.composable";
 import useMessenger from "@/message/messenger.composable";
@@ -7,6 +8,7 @@ import PendingContent from "@/spin/PendingContent.vue";
 import useCreateResource from "@/resource/createResource.composable";
 import PageTitle from "@/components/PageTitle.vue";
 import LayoutSection from "@/components/LayoutSection.vue";
+import FormKitWrapper from "@/components/FormKitWrapper.vue";
 
 const router = useRouter();
 const { requireAuthentication } = useAuth();
@@ -36,31 +38,33 @@ async function submit(fields: Form) {
   <PageTitle>{{ $t("metadata.new") }}</PageTitle>
   <LayoutSection>
     <PendingContent on="create">
-      <FormKit
-        id="create-metadata"
-        type="form"
-        :submit-label="$t('create')"
-        :submit-attrs="{
-          inputClass: 'mink-button button-primary',
-        }"
-        @submit="submit"
-      >
+      <FormKitWrapper>
         <FormKit
-          :label="$t('public_id')"
-          type="text"
-          :validation="[
-            ['required'],
-            ['trim'],
-            ['matches', /^[a-z0-9][a-z0-9-]+[a-z0-9]$/],
-          ]"
-          :validation-messages="{
-            matches: $t('metadata.public_id.error.format'),
+          id="create-metadata"
+          type="form"
+          :submit-label="$t('create')"
+          :submit-attrs="{
+            inputClass: 'mink-button button-primary',
           }"
-          name="publicId"
-          input-class="w-72 font-mono"
-          :help="$t('metadata.public_id.help')"
-        />
-      </FormKit>
+          @submit="submit"
+        >
+          <FormKit
+            :label="$t('public_id')"
+            type="text"
+            :validation="[
+              ['required'],
+              ['trim'],
+              ['matches', /^[a-z0-9][a-z0-9-]+[a-z0-9]$/],
+            ]"
+            :validation-messages="{
+              matches: $t('metadata.public_id.error.format'),
+            }"
+            name="publicId"
+            input-class="w-72 font-mono"
+            :help="$t('metadata.public_id.help')"
+          />
+        </FormKit>
+      </FormKitWrapper>
     </PendingContent>
   </LayoutSection>
 </template>