+++ CHANGELOG.md
@@ -20,6 +20,7 @@ As this project is a user-facing application, the places in the semantic version
 ### Changed
 
 - Drop unnecessary async usage of the `js-yaml` package
+- The FormKit and Highlight.js libs are imported dynamically to allow for a smaller main asset chunk
 
 ### Fixed